+++ ste/JobPartPlan.go
@@ -14,11 +14,12 @@ import (
 // dataSchemaVersion defines the data schema version of JobPart order files supported by
 // current version of azcopy
 // To be Incremented every time when we release azcopy with changed dataSchema
-const DataSchemaVersion common.Version = 15
+const DataSchemaVersion common.Version = 16
 
 const (
 	CustomHeaderMaxBytes = 256
 	MetadataMaxBytes     = 1000 // If > 65536, then jobPartPlanBlobData's MetadataLength field's type must change
+	BlobTagsMaxByte      = 4000
 	BlobTierMaxBytes     = 10
 )
 
@@ -162,7 +163,7 @@ func (jpph *JobPartPlanHeader) getString(offset int64, length int16) string {
 // TransferSrcPropertiesAndMetadata returns the SrcHTTPHeaders, properties and metadata for a transfer at given transferIndex in JobPartOrder
 // TODO: Refactor return type to an object
 func (jpph *JobPartPlanHeader) TransferSrcPropertiesAndMetadata(transferIndex uint32) (h common.ResourceHTTPHeaders, metadata common.Metadata, blobType azblob.BlobType, blobTier azblob.AccessTierType,
-	s2sGetPropertiesInBackend bool, DestLengthValidation bool, s2sSourceChangeValidation bool, s2sInvalidMetadataHandleOption common.InvalidMetadataHandleOption, entityType common.EntityType, blobVersionID string) {
+	s2sGetPropertiesInBackend bool, DestLengthValidation bool, s2sSourceChangeValidation bool, s2sInvalidMetadataHandleOption common.InvalidMetadataHandleOption, entityType common.EntityType, blobVersionID string, blobTags common.BlobTags) {
 	var err error
 	t := jpph.Transfer(transferIndex)
 
@@ -219,6 +220,11 @@ func (jpph *JobPartPlanHeader) TransferSrcPropertiesAndMetadata(transferIndex ui
 		blobVersionID = jpph.getString(offset, t.SrcBlobVersionIDLength)
 		offset += int64(t.SrcBlobVersionIDLength)
 	}
+	if t.SrcBlobTagsLength != 0 {
+		blobTagsString := jpph.getString(offset, t.SrcBlobTagsLength)
+		blobTags = common.ToCommonBlobTagsMap(blobTagsString)
+		offset += int64(t.SrcBlobTagsLength)
+	}
 	return
 }
 
@@ -271,6 +277,9 @@ type JobPartPlanDstBlob struct {
 
 	MetadataLength uint16
 	Metadata       [MetadataMaxBytes]byte
+
+	BlobTagsLength uint16
+	BlobTags       [BlobTagsMaxByte]byte
 
 	// Specifies the maximum size of block which determines the number of chunks and chunk size of a transfer
 	BlockSize int64
@@ -325,6 +334,7 @@ type JobPartPlanTransfer struct {
 	SrcBlobTypeLength           int16
 	SrcBlobTierLength           int16
 	SrcBlobVersionIDLength      int16
+	SrcBlobTagsLength           int16
 
 	// Any fields below this comment are NOT constants; they may change over as the transfer is processed.
 	// Care must be taken to read/write to these fields in a thread-safe way!