+++ ste/mgr-JobPartTransferMgr.go
@@ -20,7 +20,7 @@ import (
 type IJobPartTransferMgr interface {
 	FromTo() common.FromTo
 	Info() TransferInfo
-	ResourceDstData(dataFileToXfer []byte) (headers common.ResourceHTTPHeaders, metadata common.Metadata)
+	ResourceDstData(dataFileToXfer []byte) (headers common.ResourceHTTPHeaders, metadata common.Metadata, blobTags common.BlobTags)
 	LastModifiedTime() time.Time
 	PreserveLastModifiedTime() (time.Time, bool)
 	ShouldPutMd5() bool
@@ -86,6 +86,7 @@ type IJobPartTransferMgr interface {
 	SecurityInfoPersistenceManager() *securityInfoPersistenceManager
 	FolderDeletionManager() common.FolderDeletionManager
 	GetDestinationRoot() string
+	ShouldInferContentType() bool
 }
 
 type TransferInfo struct {
@@ -145,6 +146,7 @@ func (i TransferInfo) entityTypeLogIndicator() string {
 type SrcProperties struct {
 	SrcHTTPHeaders common.ResourceHTTPHeaders // User for S2S copy, where per transfer's src properties need be set in destination.
 	SrcMetadata    common.Metadata
+	SrcBlobTags    common.BlobTags
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -239,7 +241,7 @@ func (jptm *jobPartTransferMgr) Info() TransferInfo {
 	src, dst, _ := plan.TransferSrcDstStrings(jptm.transferIndex)
 	dstBlobData := plan.DstBlobData
 
-	srcHTTPHeaders, srcMetadata, srcBlobType, srcBlobTier, s2sGetPropertiesInBackend, DestLengthValidation, s2sSourceChangeValidation, s2sInvalidMetadataHandleOption, entityType, versionID :=
+	srcHTTPHeaders, srcMetadata, srcBlobType, srcBlobTier, s2sGetPropertiesInBackend, DestLengthValidation, s2sSourceChangeValidation, s2sInvalidMetadataHandleOption, entityType, versionID, blobTags :=
 		plan.TransferSrcPropertiesAndMetadata(jptm.transferIndex)
 	srcSAS, dstSAS := jptm.jobPartMgr.SAS()
 	// If the length of destination SAS is greater than 0
@@ -313,6 +315,16 @@ func (jptm *jobPartTransferMgr) Info() TransferInfo {
 	}
 	blockSize = common.Iffint64(blockSize > common.MaxBlockBlobBlockSize, common.MaxBlockBlobBlockSize, blockSize)
 
+	var srcBlobTags common.BlobTags
+	if blobTags != nil {
+		srcBlobTags = common.BlobTags{}
+		for k, v := range blobTags {
+			key, _ := url.QueryUnescape(k)
+			value, _ := url.QueryUnescape(v)
+			srcBlobTags[key] = value
+		}
+	}
+
 	jptm.transferInfo = &TransferInfo{
 		BlockSize:                      blockSize,
 		Source:                         src,
@@ -328,6 +340,7 @@ func (jptm *jobPartTransferMgr) Info() TransferInfo {
 		SrcProperties: SrcProperties{
 			SrcHTTPHeaders: srcHTTPHeaders,
 			SrcMetadata:    srcMetadata,
+			SrcBlobTags:    srcBlobTags,
 		},
 		SrcBlobType:    srcBlobType,
 		S2SSrcBlobTier: srcBlobTier,
@@ -428,7 +441,7 @@ func (jptm *jobPartTransferMgr) ScheduleChunks(chunkFunc chunkFunc) {
 	jptm.jobPartMgr.ScheduleChunks(chunkFunc)
 }
 
-func (jptm *jobPartTransferMgr) ResourceDstData(dataFileToXfer []byte) (headers common.ResourceHTTPHeaders, metadata common.Metadata) {
+func (jptm *jobPartTransferMgr) ResourceDstData(dataFileToXfer []byte) (headers common.ResourceHTTPHeaders, metadata common.Metadata, blobTags common.BlobTags) {
 	return jptm.jobPartMgr.(*jobPartMgr).resourceDstData(jptm.Info().Source, dataFileToXfer)
 }
 
@@ -868,4 +881,11 @@ func (jptm *jobPartTransferMgr) FolderDeletionManager() common.FolderDeletionMan
 func (jptm *jobPartTransferMgr) GetDestinationRoot() string {
 	p := jptm.jobPartMgr.Plan()
 	return string(p.DestinationRoot[:p.DestinationRootLength])
+}
+
+func (jptm *jobPartTransferMgr) ShouldInferContentType() bool {
+	// For remote files, we preserve the content-type and we don't have to infer it using AzCopy
+	// For local files, even if the file size is 0B, we try to infer the content based on file extension
+	fromTo := jptm.FromTo()
+	return fromTo.From() == common.ELocation.Local()
 }